run_command was using a wrong regular expression to check for switches
The regex /^\+|\-/ unfortunately does match as soon as the whole given options
contain any - in it, because the ^ in the given regex only counts for the +,
but not for the -. This broke our code in a place where we were overlaying
an image with another image and the given image had minuesses in it.

I added a test to explain both the problem and the fix.

Signed-off-by: Corey <probablycorey@gmail.com>