Fix Travis build for Rubinius
Installing gems with RubyGems <= 2.4.0 is not thread-safe, and because
of that Rubinius was failing when installing gems. The latest RubyGems,
2.4.1, allegedly fixes that problem, so I added

  gem update --system 2.4.1

to the Travis build. But that didn't work, the error was still there.
So I decided to just not install gems in parallel, and cache the
installed gems, so that step should still be fast enough.