Remove MiniMagick::Image#destroy!
It only deleted the tempfile in case of an error, but the main beauty of
tempfiles is that they get deleted by themselves.

Another thing is that the name was misleading. `#destroy!` sounds like
it will delete the image in any case. But that's not the case; it won't
get deleted if `MiniMagick::Image` was initialized from a permanent file
(using `MiniMagick::Image.new`).