Alias #info to #[]
It's great that we just avoid any confusion. People may use #info to
access an attribute, but they may expect that MiniMagick will fetch it
if it doesn't exist. Moreover, there is no scenario in which the user
will want to ensure that the attribute is *not* fetched when they
request it.