Output any eventual stderr from commands
The `compare` command-line tool operates in the sense that it outputs
it's info to stderr, which is very strange, but we would like to support
that. This way, people could do:

  compare = MiniMagick::Tool::Compare.new(false)

  # Do something with compare ...

  $stderr = StringIO.new
  compare.run
  output = $stderr.string
  $stderr = STDERR

It's not really pretty, but I think this is the best we can do, since
using stderr should be ultimatively fixed in ImageMagick.

Fixes #257