Make MiniMagick::Image#size return string as-is
Before the size string was turned into an integer with `#to_i`, throwing
out the size unit at the end, and converting float numbers to integers
as well. So "23KB" would become 23, "5.3MB" would become 5 etc.

This unfortunately breaks backwards compatibility, because now
MiniMagick::Image#size is now a string instead of an integer, but up
until now this value was useless anyways, so we need to make this
sacrifice.

Fixes #318