Remove automagically generated methods
The only purpose they were serving is to help MiniMagick users with
eventual typos they might make when building ImageMagick commands. If a
user makes a typo and enters an option that the ImageMagick tool doesn't
recognize, the user would get a nice NoMethodError.

However, this has brought nothing but trouble. First it introduced a
race condition in combination with autoloading, so it happens that tool
classes are ready to use before #reload_tools has finished executing,
resulting in NoMethodErrors.

The only way to avoid that was to require all of the tools upfront. The
immediate downside of this was that it prolonged the require time,
because first all methods needed to be generated, which required
executing a `<tool> -help` shell command for all of the 11 tools.
Secondly, it meant that as soon as you require MiniMagick it would have
to have the executables available. But this may not be true, maybe the
correct PATH is being set sometimes later. Anyway, people have started
experiencing these error messages after the upgrade, meaning it was a
backwards incompatible change.

This introduces a slight backwards incompatibility that now
`MiniMagick::Tool#instance_methods` won't return the list of ImageMagick
option methods anymore. But it's unlikely that someone was relying on
this behaviour, and ends justify the means.

Fixes #325