Fixed stale image info when processing
Inside of #format and #combine_options people may want to do processing
depending on the current image info. But, since info was cleared at the
beginning, referencing the info in the block caches it again, but before
the image has been processed, so after the processing we're left with
stale info. So, we clear the info only after the block has been
executed.

Fixes #330