Execute shell the same with block in Tool#call
We want to execute the shell in the same way whether the block is given
to MiniMagick::Tool#call or not, because we want the behaviour to be
consistent.

This changes the behaviour that an error will be raised if the command
fails, even if the block is passed in.