Add Image#data which uses the JSON format
The current MiniMagick::Image#details parses `identify -verbose`, and
returns it in the hash format. However, parsing this data turned out to
be very hard, it was often erroring for different kind of data.

Luckily, ImageMagick actually has the feature of returning this data in
JSON format.

  convert image.jpg json:

One notable difference from the JSON format is that the keys are in
camelcase, and there might be other subtle differences that people might
be relying on, so we create a new Minimagick::Image method for it.

Fixes #385, Fixes #379, Fixes #372