Read process output before waiting for exit
The process output is buffered, and the buffer size is limited. The
process will block on writing to stdout when the buffer is full. That's
why we need to read all data from stdout before waiting for the process
to exit.

I noticed the issue with generating a 200K image.